Add basic SSH config and only change files if they do not exist already

+++ setup.sh
@@ -1,4 +1,16 @@
 #!/bin/bash
-ln -s "$(pwd)"/bash_profile ~/.bash_profile
-ln -s "$(pwd)"/git-completion.bash ~/git-completion.bash
-. ~/.bash_profile
+
+if [ ! -f ~/.ssh/config ]; then
+	echo "Creating basic SSH config file..."
+	cp ssh-config ~/.ssh/config
+else
+	echo "Existing SSH config file found - we won't touch it."
+fi
+
+if [ ! -f ~/.bash_profile ]; then
+	echo "Symlinking bash_profile..."
+	ln -s "$(pwd)"/bash_profile ~/.bash_profile
+	. ~/.bash_profile
+else
+	echo "Existing bash_profile found - we won't touch it."
+fi
